[Hotfix] Exchange rust-xtensa repository.
Right now it is not possible to compile libcore with the xtensa
fork of the rust compiler (rev `bba6c06`).

The reasons are some flaws inside the calculation of required registers
to pass function arguments. These got triggered through a recent update
of the `compiler_builtins` helper crate which is dynamically injected
into the build process of libcore via crates.io.

As long as this is the case we must are required to use this fork which
contains a bunch of fixes.